upto Like

+++ web/views.py
@@ -1,15 +1,27 @@
+
 from django.shortcuts import render, redirect
-#from django.contrib.auth.forms import UserCreationForm 
-#from django.http import HttpResponse
 from django.contrib.auth import login, authenticate, logout
 from django.contrib.auth.models import User
-from .models import Profile, Post, Likes
+from .models import Profile, Post, Like, Comment
 from .forms import CreateBlogForm, CategoryForm, CommentForm
 from django.core.paginator import Paginator
 from django.views.decorators.http import require_POST
 from django.contrib.auth.decorators import login_required
-#from django.db import IntegrityError
 from django.views.decorators.http import require_GET
+from django.shortcuts import get_object_or_404
+
+@require_GET
+@login_required
+def add_like_comment(request, comment_id):
+    comment = get_object_or_404(Comment, id=comment_id)
+    post = comment.post  # Get the post associated with the comment
+    if Like.objects.filter(user=request.user, post=post, comment=comment).exists():
+        # If the user has already liked the comment, delete the existing like
+        Like.objects.filter(user=request.user, post=post, comment=comment).delete()
+    else:
+        # If the user hasn't liked the comment, create a new like
+        Like.objects.create(user=request.user, post=post, comment=comment)
+    return redirect("blog_detail", blog_id=comment.post.id)
 
 
 
@@ -17,68 +29,46 @@ from django.views.decorators.http import require_GET
 @require_GET
 @login_required
 def add_like(request, blog_id):
-    #blog_id = request.GET.get('blog_id')
-    try:
-        blog = Post.objects.get(id=blog_id)
-        
-    except Post.DoesNotExist:
-        return redirect("home_page")
-    else:
-        try:
-            like = Likes.objects.get(user=request.user, post=blog)
-            like.delete()
-            return redirect("blog_detail_page", blog_id=blog_id)
-        except Likes.DoesNotExist:
-            Likes.objects.create(user=request.user, post=blog)
-            return redirect("blog_detail_page", blog_id=blog_id)
-    
-#.....................................................................................................................................
+    blog = get_object_or_404(Post, id=blog_id)
+    like, created = Like.objects.get_or_create(user=request.user, post=blog)
+    if not created:
+        like.delete()
+    return redirect("blog_detail", blog_id=blog_id)
 
 @require_POST
 @login_required
 def add_comment(request):
-    #print(request.POST)
     blog_id = request.POST.get('blog_id')
-    try:
-        blog = Post.objects.get(id=blog_id)
-    except Post.DoesNotExist:
-        return redirect("home_page")
-    else:
-        form_data = CommentForm(request.POST)
-        if form_data.is_valid():
-
-            form_data.save(commenter=request.user, post=blog)
-            return redirect("blog_detail_page",blog_id=blog_id)
-         
-        else:
-            comments = blog.comments.all()
-            context = {
-                'blog':blog,
-                'comment_form':form_data,
-                'comments':comments
-
-            }
-            return render(request, 'main/blog_detail.html', context)
-
-
-#..................................................................................................
-
-def blog_detail(request, blog_id):
-    try:
-        blog = Post.objects.get(id=blog_id)
-    except Post.DoesNotExist:
-        return redirect("home_page")
+    blog = get_object_or_404(Post, id=blog_id)
+    form_data = CommentForm(request.POST)
+    if form_data.is_valid():
+        comment = form_data.save(commit=False)  # Don't save yet
+        comment.author = request.user  # Set the author field
+        comment.post = blog  # Set the post field
+        comment.save()  # Now save the comment
+        return redirect("blog_detail", blog_id=blog_id)
     else:
         comments = blog.comments.all()
-        likes = blog.likes.count
         context = {
-            'blog':blog,
-            'comment_form':CommentForm(),
-            'comments':comments,
-            'likes':likes
+            'blog': blog,
+            'content': blog.content,
+            'comment_form': form_data,
+            'comments': comments
         }
-
         return render(request, 'main/blog_detail.html', context)
+    
+
+def blog_detail(request, blog_id):
+    blog = get_object_or_404(Post, id=blog_id)
+    
+    context = {
+        'blog': blog,
+        'content': blog.content,
+        'comment_form': CommentForm(),
+        'comments': blog.comments.all()
+    }
+    return render(request, 'main/blog_detail.html', context)
+
 
 
 #.......................................................................................
@@ -125,7 +115,6 @@ def home(request):
 
 
     context={
-        #'blogs':latest_blogs
         'page_obj': page_obj
     }
     return render(request,'main/home.html', context)
@@ -189,7 +178,7 @@ def signup_view(request):
 
                 if profile_pic:
                     Profile.objects.create(profile_pic=profile_pic, user=user)
-                return redirect('login_page')
+                    return redirect("login_page")
     else:
         return render(request, 'main/signup.html')
